requirements.txt
boto3==1.14.47
Fabric3==1.13.1.post1
Flask==1.1.2
Flask-FlatPages==0.6
Flask-Script==2.0.6
Frozen-Flask==0.15
Markdown==2.5
MarkupSafe==1.1.1
Werkzeug==0.15.6
